Several big changes (VM MAZ, autoLevel changes)
Changed MAZ Special dropdown to only include map specials that your save is able to run, so won't show any options when you have a u2 hze of 1, and will gradually increase the options when you progress through u2.
Change autoLevel to use the cache that will be getting run for the fragment calculation rather than expecting the user to only be running a LMC/SMC map.
Blocked challenges from being shown in AT settings if they haven't been uploaded yet.
Added "Run Type" setting to the Void Map settings. I've removed both the Daily and C3 settings and will display the overall setting in all 3 tabs (the same as how Bone Shrine currently works) so you'll need to redo your Void Map settings.
Made a minor variable name change for Bone Shrine and have converted the setting over but there might be an issue the next time it is opened where it won't display the correct run types so double check that information is correct before saving or it might not behave the way you want it to.
so the next time you open the Bone Shrine window it won't have the correct item selected under "Run Type" so you'll need to adjust that before saving.
Bugfix for Glass dmg calc that was broken in the last update.