New mapping bugfixes
Fixed issue with Bone Shrine on shred dailies. It was only looking for if shred was active but should have also looked for if the resource you were wanting to gather was being hit by shred.
Frozen Castle will now run after Void Maps on Hypo but will run otherwise run before it.
Fix for 'MP Smithy C3' setting using the values that were in the 'MP Smithy' setting for when to run.